Revert "Bugfix/recent projects not updating"
