--- CharacterDataEditor/Constants/LogSourceConstants.cs
@@ -1,8 +1,0 @@
-﻿namespace CharacterDataEditor.Constants
-{
-    public class LogSourceConstants
-    {
-        public const string Microsoft = "Microsoft";
-        public const string System = "System";
-    }
-}
