--- CharacterDataEditor/Constants/ResourceConstants.cs
@@ -1,7 +1,0 @@
-﻿namespace CharacterDataEditor.Constants
-{
-    public class ResourceConstants
-    {
-        public const string LogoPath = "Resources/logo.png";
-    }
-}
