+++ CharacterDataEditor/GuiController.cs
@@ -1,6 +1,5 @@
 ﻿using ImGuiNET;
 using Raylib_cs;
-using System;
 using System.Numerics;
 using System.Runtime.InteropServices;
