--- CharacterDataEditor/Helpers/Win32DialogHelper.cs
@@ -1,79 +1,0 @@
-﻿using System.Runtime.InteropServices;
-using System;
-
-namespace CharacterDataEditor.Helpers
-{
-    // Code was constructed using help from https://www.pinvoke.net/default.aspx/comdlg32.GetSaveFileName
-    // We are not using System.Windows.Forms because it requires switching app to WinForms and that would
-    // disable cross-platform compatiblity.
-    // At a later time if needed, the specifics for Mac/Linux can be placed here with Preprocessor Directives used
-    // to divert code into the correct path
-
-    [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Auto)]
-    public struct OpenFileName
-    {
-        public int lStructSize;
-        public IntPtr hwndOwner;
-        public IntPtr hInstance;
-        public string lpstrFilter;
-        public string lpstrCustomFilter;
-        public int nMaxCustFilter;
-        public int nFilterIndex;
-        public string lpstrFile;
-        public int nMaxFile;
-        public string lpstrFileTitle;
-        public int nMaxFileTitle;
-        public string lpstrInitialDir;
-        public string lpstrTitle;
-        public int Flags;
-        public short nFileOffset;
-        public short nFileExtension;
-        public string lpstrDefExt;
-        public IntPtr lCustData;
-        public IntPtr lpfnHook;
-        public string lpTemplateName;
-        public IntPtr pvReserved;
-        public int dwReserved;
-        public int flagsEx;
-    }
-
-    public static class Win32DialogHelper
-    {
-        [DllImport("comdlg32.dll", SetLastError = true, CharSet = CharSet.Auto)]
-        private static extern bool GetOpenFileName(ref OpenFileName ofn);
-        [DllImport("comdlg32.dll", SetLastError = true, CharSet = CharSet.Auto)]
-        private static extern bool GetSaveFileName(ref OpenFileName ofn);
-
-        public static string ShowOpenFileDialog(string filter, string title)
-        {
-            var ofn = new OpenFileName();
-            ofn.lStructSize = Marshal.SizeOf(ofn);
-            ofn.lpstrFilter = filter;
-            ofn.lpstrFile = new string(new char[256]);
-            ofn.nMaxFile = ofn.lpstrFile.Length;
-            ofn.lpstrFileTitle = new string(new char[64]);
-            ofn.nMaxFileTitle = ofn.lpstrFileTitle.Length;
-            ofn.lpstrTitle = title;
-            if (GetOpenFileName(ref ofn))
-                return ofn.lpstrFile;
-            return string.Empty;
-        }
-
-        public static string ShowSaveFileDialog(string filter, string title, string initialPath, string defaultExtension)
-        {
-            var ofn = new OpenFileName();
-            ofn.lStructSize = Marshal.SizeOf(ofn);
-            ofn.lpstrFilter = filter;
-            ofn.lpstrFile = new string(new char[256]);
-            ofn.nMaxFile = ofn.lpstrFile.Length;
-            ofn.lpstrFileTitle = new string(new char[64]);
-            ofn.lpstrInitialDir = initialPath;
-            ofn.nMaxFileTitle = ofn.lpstrFileTitle.Length;
-            ofn.lpstrTitle = title;
-            ofn.lpstrDefExt = defaultExtension;
-            if (GetSaveFileName(ref ofn))
-                return ofn.lpstrFile;
-            return string.Empty;
-        }
-    }
-}
