--- CharacterDataEditor/Models/AttackType.cs
@@ -1,10 +1,0 @@
-﻿namespace CharacterDataEditor.Models
-{
-    public enum AttackType
-    {
-        High,
-        Mid,
-        Low,
-        Grab
-    }
-}
