--- CharacterDataEditor/Models/FrameDataModel.cs
@@ -1,8 +1,0 @@
-﻿namespace CharacterDataEditor.Models
-{
-    public record FrameDataModel
-    {
-        public int Length { get; set; }
-        public int ImageIndex { get; set; }
-    }
-}
