--- CharacterDataEditor/Models/MoveType.cs
@@ -1,22 +1,0 @@
-﻿namespace CharacterDataEditor.Models
-{
-    public enum MoveType
-    {
-        SideSpecial,
-        NeutralSpecial,
-        BackwardThrow,
-        ForwardThrow,
-        Grab,
-        JumpingHeavy,
-        JumpingMedium,
-        JumpingLight,
-        CrouchingHeavy,
-        CrouchingMedium,
-        CrouchingLight,
-        StandingHeavy,
-        StandingMedium,
-        StandingLight,
-        StandingLight2,
-        StandingLight3
-    }
-}
