--- CharacterDataEditor/Models/ProgramDataModel.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace CharacterDataEditor.Models
-{
-    public class ProgramDataModel
-    {
-        public List<RecentProjectModel> RecentlyOpenedProjects { get; set; }
-    }
-}
