--- CharacterDataEditor/Models/RGBModel.cs
@@ -1,9 +1,0 @@
-﻿namespace CharacterDataEditor.Models
-{
-    public record RGBModel
-    {
-        public int Red { get; set; }
-        public int Blue { get; set; }
-        public int Green { get; set; }
-    }
-}
