--- CharacterDataEditor/Options/ArgValues.cs
@@ -1,8 +1,0 @@
-﻿namespace CharacterDataEditor.Options
-{
-    public class ArgValues
-    {
-        public string LogPath { get; set; }
-        public bool EnableConsole { get; set; } = false;
-    }
-}
