--- CharacterDataEditor/Startup.cs
@@ -1,25 +1,0 @@
-﻿using CharacterDataEditor.Extensions;
-using CharacterDataEditor.Services;
-using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.Collections;
-
-namespace CharacterDataEditor
-{
-    public class Startup
-    {
-        public Startup()
-        {
-        }
-
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddSingleton<IRenderUI, RaylibImGui>();
-            services.AddSingleton<IRecentFiles, RecentFiles>();
-            services.AddSingleton<ICharacterOperations, CharacterOperations>();
-
-            //register screens
-            services.RegisterScreens();
-        }
-    }
-}
